Moved int tests fixtures to conftest
